feat(android-backend): refactor module
